remove auth duplicate
